Don't make an error when we generate a kymograph twice.
Before this commit, there was an error when we tried to generate a
kymograph after the a first from the TM action. The tracing tool
tried to register a second time to the IJ toolbar, which generated
an error showed in the IJ log.
The changes introduced by this commit emulated what was done for
TrackMate spot edit tool:
The tracing tool is not a singleton instance, that has a map of
ImagePlus to tracing objects (in a Bundle).